# PERF: read_html: Reading one HTML file with multiple tables is much slower than loading each table separatly