--- a/pandas/io/html.py
+++ b/pandas/io/html.py
@@ -744,6 +744,7 @@ class _LxmlFrameParser(_HtmlFrameParser):
         pattern = match.pattern
 
         # 1. check all descendants for the given pattern and only search tables
+        # GH 49929
         xpath_expr = f"//table[.//text()[re:test(., {repr(pattern)})]]"
 
         # if any table attributes were given build an xpath expression to